# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UsageEntry": {
        "properties": {
          "faithfulness_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faithfulness Score"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "log_id": {
            "title": "Log Id",
            "type": "integer"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "tokens_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Used"
          },
          "total_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost Usd"
          }
        },
        "required": [
          "log_id",
          "timestamp",
          "query"
        ],
        "title": "UsageEntry",
        "type": "object"
      },
      "UsageResponse": {
        "properties": {
          "queries": {
            "items": {
              "$ref": "#/components/schemas/UsageEntry"
            },
            "title": "Queries",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/UsageSummary"
          }
        },
        "required": [
          "summary",
          "queries"
        ],
        "title": "UsageResponse",
        "type": "object"
      },
      "UsageSummary": {
        "properties": {
          "avg_cost_per_query_usd": {
            "title": "Avg Cost Per Query Usd",
            "type": "number"
          },
          "avg_latency_ms": {
            "title": "Avg Latency Ms",
            "type": "number"
          },
          "total_cost_usd": {
            "title": "Total Cost Usd",
            "type": "number"
          },
          "total_queries": {
            "title": "Total Queries",
            "type": "integer"
          }
        },
        "required": [
          "total_queries",
          "total_cost_usd",
          "avg_cost_per_query_usd",
          "avg_latency_ms"
        ],
        "title": "UsageSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceRetrieval": {
        "description": "The retrieval settings a workspace's chat runs. Anything left out keeps the deployment's default.",
        "properties": {
          "dense_weight": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dense Weight"
          },
          "rerank_candidates": {
            "anyOf": [
              {
                "maximum": 80.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Candidates"
          },
          "reranker": {
            "anyOf": [
              {
                "enum": [
                  "flashrank",
                  "jina",
                  "voyage"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reranker"
          },
          "rrf_k": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rrf K"
          },
          "sparse_weight": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sparse Weight"
          },
          "strategy": {
            "anyOf": [
              {
                "enum": [
                  "dense",
                  "sparse",
                  "hybrid"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          }
        },
        "title": "WorkspaceRetrieval",
        "type": "object"
      },
      "WorkspaceSettingsResponse": {
        "properties": {
          "retrieval": {
            "$ref": "#/components/schemas/WorkspaceRetrieval"
          },
          "source": {
            "enum": [
              "workspace",
              "default"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "retrieval"
        ],
        "title": "WorkspaceSettingsResponse",
        "type": "object"
      },
      "WorkspaceStats": {
        "properties": {
          "documents_count": {
            "title": "Documents Count",
            "type": "integer"
          },
          "total_chunks": {
            "title": "Total Chunks",
            "type": "integer"
          },
          "total_tokens": {
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "required": [
          "documents_count",
          "total_chunks",
          "total_tokens"
        ],
        "title": "WorkspaceStats",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Admin-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Nexus: retrieval-augmented question answering over your own documents, with the tools to measure how well it retrieves. Authenticate with `X-API-Key` (a workspace key); `/v1/admin` takes `X-Admin-Key`. Every error has the body `{code, message, request_id}`.",
    "title": "Nexus RAG API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/usage": {
      "get": {
        "description": "Cost and latency of this workspace's questions: totals over its retained history, and the most recent ones.",
        "operationId": "usage",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Usage",
        "tags": [
          "workspace"
        ]
      }
    },
    "/v1/workspace/settings": {
      "delete": {
        "description": "Returns this workspace to the deployment's defaults.",
        "operationId": "reset_workspace_settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Reset Workspace Settings",
        "tags": [
          "workspace"
        ]
      },
      "get": {
        "description": "The retrieval settings chat runs for this workspace, and whether they are its own or the deployment's defaults.",
        "operationId": "get_workspace_settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Workspace Settings",
        "tags": [
          "workspace"
        ]
      },
      "put": {
        "description": "Sets the retrieval settings this workspace's chat runs (for example the winner of an experiment).",
        "operationId": "put_workspace_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceRetrieval"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Put Workspace Settings",
        "tags": [
          "workspace"
        ]
      }
    },
    "/v1/workspace/stats": {
      "get": {
        "description": "Documents, chunks and tokens in the caller's workspace (every workspace for the admin).",
        "operationId": "workspace_stats",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Workspace Stats",
        "tags": [
          "workspace"
        ]
      }
    }
  }
}